[feat/#83] 화면 이동간 애니메이션 추가
