+++ app/src/main/java/com/teumteum/teumteum/presentation/group/join/check/GroupMeetCheckActivity.kt
@@ -1,4 +1,4 @@
-package com.teumteum.teumteum.presentation.group.join
+package com.teumteum.teumteum.presentation.group.join.check
 
 import android.content.Context
 import android.content.Intent
@@ -10,22 +10,21 @@ import com.teumteum.base.BindingActivity
 import com.teumteum.base.component.appbar.AppBarLayout
 import com.teumteum.base.component.appbar.AppBarMenu
 import com.teumteum.base.databinding.LayoutCommonAppbarBinding
-import com.teumteum.base.util.extension.defaultSnackBar
+import com.teumteum.base.util.extension.defaultToast
 import com.teumteum.base.util.extension.longExtra
-import com.teumteum.base.util.extension.toast
 import com.teumteum.teumteum.R
-import com.teumteum.teumteum.databinding.ActivityGroupMeetCompleteBinding
-import com.teumteum.teumteum.presentation.MainActivity
+import com.teumteum.teumteum.databinding.ActivityGroupMeetCheckBinding
+import com.teumteum.teumteum.presentation.group.join.complete.GroupMeetCompleteActivity
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 
 @AndroidEntryPoint
-class GroupMeetCompleteActivity :
-    BindingActivity<ActivityGroupMeetCompleteBinding>(R.layout.activity_group_meet_complete),
+class GroupMeetCheckActivity :
+    BindingActivity<ActivityGroupMeetCheckBinding>(R.layout.activity_group_meet_check),
     AppBarLayout {
 
-    private val viewModel by viewModels<GroupMeetCompleteViewModel>()
+    private val viewModel by viewModels<GroupMeetCheckViewModel>()
     private val meetingId by longExtra()
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -48,23 +47,32 @@ class GroupMeetCompleteActivity :
                 useRippleEffect = false,
                 clickEvent = {
                     finish()
+                    closeActivitySlideAnimation()
                 }
             )
         )
     }
 
+    private fun initEvent() {
+        binding.checkboxWaring.setOnCheckedChangeListener { _, isCheck ->
+            binding.btnJoin.isEnabled = isCheck
+        }
+
+        binding.btnJoin.setOnClickListener {
+            viewModel.joinGroup(meetingId)
+        }
+    }
+
     private fun observe() {
-        viewModel.cancelState.flowWithLifecycle(lifecycle)
+        viewModel.joinState.flowWithLifecycle(lifecycle)
             .onEach {
                 when (it) {
-                    is GroupMeetCancelUiState.Success -> {
-                        defaultSnackBar(binding.root, getString(R.string.group_meet_cancel))
-                        goToMainActivity()
-                        finish()
+                    is MeetCheckUiState.Success -> {
+                        startActivity(GroupMeetCompleteActivity.getIntent(this, it.data.id))
+                        openActivitySlideAnimation()
                     }
-
-                    is GroupMeetCancelUiState.Failure -> {
-                        defaultSnackBar(binding.root, it.msg)
+                    is MeetCheckUiState.Failure -> {
+                        defaultToast(it.msg)
                     }
 
                     else -> {}
@@ -72,27 +80,9 @@ class GroupMeetCompleteActivity :
             }.launchIn(lifecycleScope)
     }
 
-    private fun initEvent() {
-        binding.btnCancel.setOnClickListener {
-            viewModel.deleteJoinCancel(meetingId)
-        }
-
-        binding.btnHome.setOnClickListener {
-            goToMainActivity()
-            finish()
-        }
-    }
-
-    private fun goToMainActivity() {
-        Intent(this, MainActivity::class.java).apply {
-            flags = Intent.FLAG_ACTIVITY_CLEAR_TOP
-            startActivity(this)
-        }
-    }
-
     companion object {
         fun getIntent(context: Context, meetingId: Long) =
-            Intent(context, GroupMeetCompleteActivity::class.java).apply {
+            Intent(context, GroupMeetCheckActivity::class.java).apply {
                 putExtra("meetingId", meetingId)
             }
     }
