+++ app/src/main/java/com/teumteum/teumteum/presentation/moim/MoimConfirm.kt
@@ -1,8 +1,6 @@
 package com.teumteum.teumteum.presentation.moim
 
 import android.app.Activity
-import android.content.Context
-import android.util.Log
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -42,10 +40,10 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.clipToBounds
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
 import coil.compose.rememberAsyncImagePainter
 import com.google.accompanist.pager.ExperimentalPagerApi
 import com.google.accompanist.pager.rememberPagerState
+import com.teumteum.base.BindingActivity
 import com.teumteum.base.component.compose.TeumDivider
 import com.teumteum.base.component.compose.TeumDividerHorizontalThick
 import com.teumteum.base.component.compose.TeumDividerThick
@@ -58,7 +56,10 @@ import com.teumteum.base.component.compose.theme.TmtmColorPalette
 import com.teumteum.base.util.extension.toast
 import com.teumteum.domain.entity.Friend
 import com.teumteum.teumteum.R
-import com.teumteum.teumteum.presentation.group.join.GroupMeetCheckActivity
+import com.teumteum.teumteum.presentation.group.join.JoinFriendListActivity
+import com.teumteum.teumteum.presentation.group.join.check.GroupMeetCheckActivity
+import kotlinx.serialization.encodeToString
+import kotlinx.serialization.json.Json
 
 
 @OptIn(ExperimentalPagerApi::class)
@@ -66,7 +67,10 @@ import com.teumteum.teumteum.presentation.group.join.GroupMeetCheckActivity
 fun MoimConfirm(viewModel: MoimViewModel, activity: Activity, isJoinView: Boolean) {
     TmScaffold(
         topbarText = if (isJoinView) "" else stringResource(id = R.string.moim_confirm_appbar),
-        onClick = { activity.finish() }
+        onClick = {
+            activity.finish()
+            (activity as? BindingActivity<*>)?.closeActivitySlideAnimation()
+        }
     ) {
         val scrollState = rememberScrollState()
         Column(
@@ -86,7 +90,8 @@ fun MoimConfirm(viewModel: MoimViewModel, activity: Activity, isJoinView: Boolea
                 TeumDividerHorizontalThick(int = 1, 20)
                 TmMarginVerticalSpacer(size = 20)
                 MoimJoinUserRow(viewModel) {
-                    activity.toast("it")
+                    activity.startActivity(JoinFriendListActivity.getIntent(activity, Json.encodeToString(it)))
+                    (activity as? BindingActivity<*>)?.openActivitySlideAnimation()
                 }
                 TmMarginVerticalSpacer(size = 20)
             }
@@ -95,7 +100,10 @@ fun MoimConfirm(viewModel: MoimViewModel, activity: Activity, isJoinView: Boolea
             Spacer(modifier = Modifier.weight(1f))
             TeumDivider()
             if (isJoinView) {
-                MoimJoinBtn(viewModel = viewModel) { activity.startActivity(GroupMeetCheckActivity.getIntent(activity, it)) }
+                MoimJoinBtn(viewModel = viewModel) {
+                    activity.startActivity(GroupMeetCheckActivity.getIntent(activity, it))
+                    (activity as? BindingActivity<*>)?.openActivitySlideAnimation()
+                }
             } else {
                 MoimCreateBtn(text = stringResource(id = R.string.moim_next_btn), viewModel = viewModel)
                 TmMarginVerticalSpacer(size = 24)
